introducing capsule mixins
