+++ robottelo/host_helpers/__init__.py
@@ -1,3 +1,4 @@
+from robottelo.host_helpers.capsule_mixins import CapsuleInfo
 from robottelo.host_helpers.contenthost_mixins import SystemFacts
 from robottelo.host_helpers.contenthost_mixins import VersionedContent
 from robottelo.host_helpers.satellite_mixins import ContentInfo
@@ -5,6 +6,10 @@ from robottelo.host_helpers.satellite_mixins import Factories
 
 
 class ContentHostMixins(SystemFacts, VersionedContent):
+    pass
+
+
+class CapsuleMixins(CapsuleInfo):
     pass
